Membrain-seg should generate binned mrc itself
